Refactor Improve strict mode expression handling and validation
