+++ fixture-monkey/src/main/java/com/navercorp/fixturemonkey/FixtureMonkeyBuilder.java
@@ -61,14 +61,10 @@ import com.navercorp.fixturemonkey.buildergroup.ArbitraryBuilderCandidate;
 import com.navercorp.fixturemonkey.buildergroup.ArbitraryBuilderGroup;
 import com.navercorp.fixturemonkey.customizer.MonkeyManipulatorFactory;
 import com.navercorp.fixturemonkey.expression.ArbitraryExpressionFactory;
-import com.navercorp.fixturemonkey.expression.MonkeyExpression;
 import com.navercorp.fixturemonkey.expression.MonkeyExpressionFactory;
-import com.navercorp.fixturemonkey.expression.StrictModeNextNodePredicateList;
+import com.navercorp.fixturemonkey.expression.StrictModeMonkeyExpressionFactory;
 import com.navercorp.fixturemonkey.resolver.ManipulatorOptimizer;
 import com.navercorp.fixturemonkey.resolver.NoneManipulatorOptimizer;
-import com.navercorp.fixturemonkey.tree.ApplyStrictModeResolver;
-import com.navercorp.fixturemonkey.tree.NextNodePredicate;
-import com.navercorp.fixturemonkey.tree.NodeResolver;
 
 @SuppressWarnings("unused")
 @API(since = "0.4.0", status = Status.MAINTAINED)
@@ -499,34 +495,7 @@ public final class FixtureMonkeyBuilder {
 	}
 
 	public FixtureMonkeyBuilder useExpressionStrictMode() {
-		this.monkeyExpressionFactory = new MonkeyExpressionFactory() {
-			private final MonkeyExpressionFactory delegate = new ArbitraryExpressionFactory();
-
-			@Override
-			public MonkeyExpression from(String expression) {
-				return from(expression, null);
-			}
-
-			@Override
-			public MonkeyExpression from(String expression, Class<?> rootClass) {
-				MonkeyExpression monkeyExpression = delegate.from(expression);
-				return new MonkeyExpression() {
-					@Override
-					public NodeResolver toNodeResolver() {
-						return new ApplyStrictModeResolver(monkeyExpression.toNodeResolver());
-					}
-
-					@Override
-					public List<NextNodePredicate> toNextNodePredicate() {
-						if (rootClass == null) {
-							return monkeyExpression.toNextNodePredicate();
-						}
-						List<NextNodePredicate> predicates = monkeyExpression.toNextNodePredicate();
-						return new StrictModeNextNodePredicateList(predicates, rootClass);
-					}
-				};
-			}
-		};
+		this.monkeyExpressionFactory = new StrictModeMonkeyExpressionFactory(new ArbitraryExpressionFactory());
 		return this;
 	}
