+++ fixture-monkey/src/main/java/com/navercorp/fixturemonkey/expression/StrictModeMonkeyExpressionFactory.java
@@ -1,0 +1,68 @@
+/*
+ * Fixture Monkey
+ *
+ * Copyright (c) 2021-present NAVER Corp.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.navercorp.fixturemonkey.expression;
+
+import java.util.List;
+
+import com.navercorp.fixturemonkey.tree.ApplyStrictModeResolver;
+import com.navercorp.fixturemonkey.tree.NextNodePredicate;
+import com.navercorp.fixturemonkey.tree.NodeResolver;
+
+public class StrictModeMonkeyExpressionFactory implements MonkeyExpressionFactory {
+	private final MonkeyExpressionFactory delegate;
+
+	public StrictModeMonkeyExpressionFactory(MonkeyExpressionFactory delegate) {
+		this.delegate = delegate;
+	}
+
+	@Override
+	public MonkeyExpression from(String expression) {
+		return from(expression, null);
+	}
+
+	@Override
+	public MonkeyExpression from(String expression, Class<?> rootClass) {
+		MonkeyExpression monkeyExpression = delegate.from(expression);
+		return new StrictModeMonkeyExpression(monkeyExpression, rootClass);
+	}
+
+	private static class StrictModeMonkeyExpression implements MonkeyExpression {
+		private final MonkeyExpression delegate;
+		private final Class<?> rootClass;
+
+		public StrictModeMonkeyExpression(MonkeyExpression delegate, Class<?> rootClass) {
+			this.delegate = delegate;
+			this.rootClass = rootClass;
+		}
+
+		@Override
+		public NodeResolver toNodeResolver() {
+			return new ApplyStrictModeResolver(delegate.toNodeResolver());
+		}
+
+		@Override
+		public List<NextNodePredicate> toNextNodePredicate() {
+			if (rootClass == null) {
+				return delegate.toNextNodePredicate();
+			}
+			List<NextNodePredicate> predicates = delegate.toNextNodePredicate();
+			return new StrictModeNextNodePredicateContainer(predicates, rootClass);
+		}
+	}
+}
